Refactoring: inlined 'attack' index calculator
